refactor: update api 엔드포인트 변경
